+++ packages/camera/camera_android/lib/src/utils.dart
@@ -17,6 +17,32 @@ CameraLensDirection cameraLensDirectionFromPlatform(
   };
 }
 
+/// Converts a [PlatformDeviceOrientation] to [DeviceOrientation].
+DeviceOrientation deviceOrientationFromPlatform(
+        PlatformDeviceOrientation orientation) =>
+    switch (orientation) {
+      PlatformDeviceOrientation.portraitUp => DeviceOrientation.portraitUp,
+      PlatformDeviceOrientation.portraitDown => DeviceOrientation.portraitDown,
+      PlatformDeviceOrientation.landscapeLeft =>
+        DeviceOrientation.landscapeLeft,
+      PlatformDeviceOrientation.landscapeRight =>
+        DeviceOrientation.landscapeRight,
+    };
+
+/// Converts a [PlatformExposureMode] to [ExposureMode].
+ExposureMode exposureModeFromPlatform(PlatformExposureMode exposureMode) =>
+    switch (exposureMode) {
+      PlatformExposureMode.auto => ExposureMode.auto,
+      PlatformExposureMode.locked => ExposureMode.locked,
+    };
+
+/// Converts a [PlatformFocusMode] to [FocusMode].
+FocusMode focusModeFromPlatform(PlatformFocusMode focusMode) =>
+    switch (focusMode) {
+      PlatformFocusMode.auto => FocusMode.auto,
+      PlatformFocusMode.locked => FocusMode.locked,
+    };
+
 /// Returns the device orientation as a String.
 String serializeDeviceOrientation(DeviceOrientation orientation) {
   switch (orientation) {
@@ -36,20 +62,4 @@ String serializeDeviceOrientation(DeviceOrientation orientation) {
   // switch as needing an update.
   // ignore: dead_code
   return 'portraitUp';
-}
-
-/// Returns the device orientation for a given String.
-DeviceOrientation deserializeDeviceOrientation(String str) {
-  switch (str) {
-    case 'portraitUp':
-      return DeviceOrientation.portraitUp;
-    case 'portraitDown':
-      return DeviceOrientation.portraitDown;
-    case 'landscapeRight':
-      return DeviceOrientation.landscapeRight;
-    case 'landscapeLeft':
-      return DeviceOrientation.landscapeLeft;
-    default:
-      throw ArgumentError('"$str" is not a valid DeviceOrientation value');
-  }
 }